Update the instructions about registering on PyPI
As the process is complicated and complex, and will definitely take a
lot of time,
it seems more appropriate and less error-prone to only go through this on
the test PyPI instance. The students can get back to the materials to
learn how to update their configurations for the real PyPI; it's just
skipped in the main lesson body.